--- src/api/index.js
@@ -1,1 +1,0 @@
-export { signUp } from './signUp';
