+++ src/hooks/index.js
@@ -1,1 +1,2 @@
 export { default as useModal } from './useModal';
+export { default as useSignUpAPI } from './useSignUpAPI';
