+++ src/utils/index.js
@@ -1,15 +1,2 @@
-export const getLocalStorage = (key) => {
-  try {
-    const item = localStorage.getItem(key);
-
-    return JSON.parse(item);
-  } catch (error) {
-    console.error(error);
-
-    return undefined;
-  }
-};
-
-export const setLocalStorage = (key, value) => {
-  return localStorage.setItem(key, JSON.stringify(value));
-};
+export { default as request } from './request';
+export { getLocalStorage, setLocalStorage } from './localStorage';
