+++ src/utils/localStorage.js
@@ -1,0 +1,15 @@
+export const getLocalStorage = (key) => {
+  try {
+    const item = localStorage.getItem(key);
+
+    return JSON.parse(item);
+  } catch (error) {
+    console.error(error);
+
+    return undefined;
+  }
+};
+
+export const setLocalStorage = (key, value) => {
+  return localStorage.setItem(key, JSON.stringify(value));
+};
